make max collection size configurable: the default 255 sometimes is too big for bigger objects having a lot of lists

+++ random-beans/src/main/java/io/github/benas/randombeans/util/Constants.java
@@ -44,7 +44,7 @@ public abstract class Constants {
     /**
      * Maximum collection size.
      */
-    public static final byte MAX_COLLECTION_SIZE = Byte.MAX_VALUE;
+    public static final int MAX_COLLECTION_SIZE = Byte.MAX_VALUE;
 
     /**
      * Maximum string size.
